improvment: mejoro practicas separando tareas en carpetas para tener un codigo mas limpio
